requirements.txt
# PDF处理
pdf2image>=1.16.3,<2.0.0    # PDF转图片
PyMuPDF>=1.23.8,<2.0.0      # PDF文本提取

# Word文档处理
python-docx>=0.8.11,<1.0.0  # Word文档处理

# OCR相关
pytesseract>=0.3.10,<0.4.0  # OCR文字识别
Pillow>=10.0.0,<11.0.0      # 图片处理

# HTTP请求
requests>=2.31.0,<3.0.0     # API调用

# 类型提示
typing-extensions>=4.8.0,<5.0.0  # Python类型提示扩展

# 路径处理
pathlib>=1.0.1,<2.0.0       # 路径处理工具 

# 语言检测
langdetect>=1.0.9,<2.0.0    # 语言检测 

# Web框架
flask>=2.0.0,<3.0.0         # Web应用框架
flask-cors>=4.0.0,<5.0.0    # 跨域支持
werkzeug>=2.0.0,<3.0.0      # WSGI工具库 

# 新增依赖
flask-socketio>=5.1.1    # WebSocket支持
redis>=4.0.0            # Redis缓存
pandas>=1.3.0           # 电子表格处理
python-pptx>=0.6.21     # PPT处理
ebooklib>=0.17.1        # 电子书处理
beautifulsoup4>=4.9.3   # HTML解析 

# 核心依赖
pdf2image>=1.16.3
PyMuPDF>=1.23.8
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
typing-extensions>=4.8.0
pathlib>=1.0.1
langdetect>=1.0.9

# Web框架
flask>=2.0.0
flask-socketio>=5.1.1
flask-cors>=4.0.0
werkzeug>=2.0.0

# 数据处理
redis>=4.0.0
pandas>=1.3.0
python-pptx>=0.6.21
ebooklib>=0.17.1
beautifulsoup4>=4.9.3
markdown>=3.4.0 

# 图像处理
opencv-python>=4.8.0    # 图像处理和倾斜校正
numpy>=1.24.0          # 数值计算 